Silo progress support into `progress.rs`
